refactor(install): P1.L2 -- extract local_content.py phase module
Moves local-content leaf helpers (root project's .apm/ as implicit
local package per #714, local-path dependencies from apm.yml) into
`apm_cli/install/phases/local_content.py`.  Three of the four planned
functions form a coherent feature with no coupling to the integration
pipeline -- they take a project root and return booleans or copy files.

Functions moved:
- _project_has_root_primitives
- _has_local_apm_content
- _copy_local_package

Function KEPT in commands/install.py:
- _integrate_local_content  (calls _integrate_package_primitives via
  bare-name lookup; 5 tests patch
  apm_cli.commands.install._integrate_package_primitives to intercept
  that call -- same L1 lesson as _validate_and_add_packages_to_apm_yml)

`commands/install.py` re-exports the three moved names so existing
callers (_install_apm_dependencies) and any future @patch targets
keep working.  Targeted install tests (196) + full unit suite (3972)
green.

install.py LOC: 2630 -> 2562.
